requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.1.0
dash-html-components==2.0.0
dash-table==5.0.0
pandas==2.1.0
numpy==1.24.3
plotly==5.18.0
scikit-learn==1.3.0
google-cloud-bigquery==3.15.0
google-cloud-storage==2.15.0
vertexai==0.24.0
python-dotenv==1.0.0
nltk==3.8.1
